Disable backtrace capturing in Embedded Swift. (#1339)
This PR disables the bulk of the backtrace-capturing-on-error logic when
building for Embedded Swift. At least for the moment, we can't deal with
error existentials in Embedded Swift, so this code won't function
correctly even if it successfully compiles.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.