Unconditionally conform `Attachment` to `CustomStringConvertible`.
This PR adjusts `Attachment`'s conformance to `CustomStringConvertible`
such that it conforms even when `AttachableValue` is not `Copyable`. We
now check at runtime (by way of protocol shenanigans, of course) whether
the attachable value is copyable and, if so, take a different code path
than we take if it does not conform.