Add a property to `Test.Case` to get the values of its arguments.
This PR adds an experimental property to `Test.Case` to allow callers to get the
test case's arguments as a type-erased array. For example:

```swift
if let testCase = Test.Case.current {
  for value in testCase.argumentValues {
    print("\(value): \(type(of: value))")
  }
}
```

We already have an `arguments` property of a type that we do not intend to make
API, so the new property is named `argumentValues` pending any of SPI renaming.