Attachment lifetimes (#1319)
This PR introduces a new experimental trait, `.savingAttachments(if:)`,
that can be used to control whether a test's attachments are saved or
not.

XCTest has API around the
[`XCTAttachment.Lifetime`](https://developer.apple.com/documentation/xctest/xctattachment/lifetime-swift.enum)
enumeration that developers can use to control whether attachments are
saved to a test report in Xcode. This enumeration has two cases:

```objc
/*
 * Attachment will be kept regardless of the outcome of the test.
 */
XCTAttachmentLifetimeKeepAlways = 0,

/*
 * Attachment will only be kept when the test fails, and deleted otherwise.
 */
XCTAttachmentLifetimeDeleteOnSuccess = 1
```

I've opted to implement something a bit more granular. A developer can
specify `.savingAttachments(if: .testFails)` and `.savingAttachments(if:
.testPasses)` or can call some custom function of their own design like
`runningInCI` or `hasPlentyOfFloppyDiskSpace`. The default behaviour if
this trait is not used is to always save attachments, which is
equivalent to `XCTAttachmentLifetimeKeepAlways`.
`XCTAttachmentLifetimeDeleteOnSuccess` is, in effect, equivalent to
`.savingAttachments(if: .testFails)`, but I hope reads a bit more
clearly in context.

Here's a usage example:

```swift
@Test(.savingAttachments(if: .testFails))
func `best test ever`() {
  Attachment.record("...") // only saves to the test report or to disk if the
                           // next line is uncommented.
  // Issue.record("sadness")
}
```

I've taken the opportunity to update existing documentation for
`Attachment` and `Attachable` to try to use more consistent language: a
test records an attachment and then the testing library saves it
(somewhere). I'm sure I've missed some spots, so please point them out
if you see them.

Resolves rdar://138921461.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.