Make `Attachment` always conform to `CustomStringConvertible`.
This PR ensures that `Attachment` conforms to `CustomStringConvertible` even if
the value it wraps is non-copyable. We accomplish this by creating a new helper
function that is able to box an arbitrary `~Copyable` value in `Any` if its type
is actually `Copyable`. (This requires a newer runtime on Apple platforms, so we
fall back to a generic (`~Copyable`) implementation on older macOS/iOS/etc.) We
then call this helper function in `Attachment.description` and if it returns a
value, we can stringify it.