Create fallback event handler library
Minimal version of the changes in Jonathan's original PR here:
https://github.com/swiftlang/swift-testing/pull/1280

The new library is renamed to be "_TestingInterop" to more closely match
its intended (and limited) purpose, but that's open to further
discussion.

This library will eventually be visible in the toolchain, and either
testing library (Swift Testing or XCTest) will be able to use it to pass
along unhandled issues to a test runner from the other framework,
enabling interoperability.