Capture the current working directory early on OpenBSD.
This PR adds some OpenBSD-specific code to capture the current working directory as early as possible. We then use that directory when argv[0] appears to be a relative path so that the following incantation works correctly:

```sh
./.build/debug/myPackageTests.xctest --testing-library swift-testing
```

This logic is not necessary on other platforms because they all provide a way to get the path to the current executable. OpenBSD has no such API.