Update the comment and bug ID explaining why a workaround in ConfirmationTests is still needed (#1391)
This has no functional change, it just updates the comment and tracking
bug ID explaining why a certain workaround in `ConfirmationTests.swift`
for existential composition of parameterized protocol types is still
necessary.

### Motivation:

I tried removing this workaround recently, since it should no longer be
necessary, but discovered that while the syntax is now valid in Swift
6.2, there's a new blocker—the runtime crash tracked by the new bug ID.
It's helpful to document this so we know why the workaround is still
needed.

Note that even if/when this remaining issue is resolved, this capability
will require a certain minimum OS version on Apple platforms, since the
compiler fix had a runtime component.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.