Avoid `readToEnd()` on pipes.
Attempting to read to the end of a pipe will block if the writer hasn't finished
writing bytes. This can lead to a deadlock (or, at least, pathologically bad
read performance) if the writer is _also_ communicating with the reader via
another file descriptor or other IPC mechanism. It can also lead to
difficult-to-describe issues when the pipe's kernel-side buffer is filled and
the writer is blocked waiting for the reader to empty it.

This PR introduces a function on `FileHandle` that reads line-by-line which
gives the reader a chance to process data as it comes in rather than waiting for
EOF.