Capture the current working directory early on OpenBSD. (#1388)
This PR adds some OpenBSD-specific code to capture the current working
directory as early as possible. We then use that directory when argv[0]
appears to be a relative path so that the following incantation works
correctly:

```sh
./.build/debug/myPackageTests.xctest --testing-library swift-testing
```

This logic is not necessary on other platforms because they all provide
a way to get the path to the current executable. OpenBSD has no such
API.

Resolves #1348.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.