Avoid a potential race condition and use-after-free when calling `Test.cancel()`.
If a test creates an unstructured, non-detached task that continues running
after the test has finished and eventually calls `Test.cancel()`, then it may be
able to see a reference to the test's (or test case's) task after it has been
destroyed by the Swift runtime.

This PR ensures that the infrastructure under `Test.cancel()` clears its
reference to the test's task before returning. This then minimizes the risk of
observing the task after it has been destroyed.

Further work at the Swift runtime level may be required to completely eliminate
this race condition, but this change makes it sufficiently narrow that any
example I can come up with is contrived.