Fix a regression that prevents saving attachments.
This PR ensures that the attachment-saving logic in `Runner` is applied
consistently. I made the mistake of moving that logic to
`configureEventHandlerRuntimeState()` because I'd forgotten that that
function only has an effect when we're running our own tests. And, as a
result, all our tests continued to pass because the code was in place
for our tests (and for nobody else.)

I've moved the code to a different location that we will always call
when running a `Runner` instance, so the issue is resolved now. I don't
have a great way to set up a unit test for this; tested manually at
desk.