Handle image types derived from the base types. (#1399)
On Apple platforms, you can declare file types (uniform type
identifiers, or "UTTypes" for short) that conform to existing types.
That means a developer could conceivably say "save this image as .florb"
where `.florb` represents a type that conforms to, but is not identical
to, `.jpeg`.

This PR ensures that if a developer passes such a type, we handle it by
passing the supported base type to Image I/O. And, if the developer
passes some image format that is unsupported by Image I/O, we can throw
an error with a more specific diagnostic than just "it failed."

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.