When saving a `String` as an attachment, assume `".txt"` as a path extension.
When a string is attached to a test, we encode it as UTF-8 (as is tradition).
This PR adjusts `String`'s conformance to `Attachable` so that it uses the
`".txt"` extension (unless the user specifies something else). This improves
usability in Xcode test reports by making these attachments double-clickable
and QuickLookable (is that a word?)