When saving a `String` as an attachment, assume `".txt"` as a path extension. (#1404)
When a string is attached to a test, we encode it as UTF-8 (as is
tradition). This PR adjusts `String`'s conformance to `Attachable` so
that it uses the `".txt"` extension (unless the user specifies something
else). This improves usability in Xcode test reports by making these
attachments double-clickable and QuickLookable (is that a word?)

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.