Add qualification status for supported platforms in the README
We've been able to build for more platforms in our CI, so update the
README to reflect this triumph. Explain how the different qualification
levels differ, and focus on build + test for all platforms in the future.

Move some platforms from Experimental -> Supported, although not with
full CI build + test yet.