Ignore ExpectationFailedError when thrown by the body closure passed to withKnownIssue (#1441)
This fixes an issue where a redundant, second known issue is recorded
when `try #require()` fails within the body closure of `withKnownIssue`.
For example, in the following example code:

```swift
@Test func example() {
  withKnownIssue {
    try #require(Bool(false))
  }
}
```

two (known) issues are recorded:

> ✘ Test example() recorded a known issue at
KnownIssueTests.swift:641:9: Expectation failed: Bool(false)
> ✘ Test example() recorded a known issue at
KnownIssueTests.swift:640:18: An API was misused
> ↳ Recorded an error of type ExpectationFailedError representing an
expectation that failed and was already recorded: ...

Of these, only the first should be recorded and the second should be
suppressed.

Fixes rdar://153550847

### Motivation:

It should be possible to use `try #require()` in the body closure passed
to `withKnownIssue`, and _matching_ failures should only cause a single
known issue to be recorded. The second issue shown above describes an
infrastructural `Error`-conforming type which is not generally intended
to be included in the results test authors see, and can cause confusion.

### Modifications:

- When attempting to match an error thrown by a `withKnownIssue` body
closure, ignore the error if it's of type `ExpectationFailedError.
- Add a new unit test validating this, which fails prior to the fix.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.