Bump `ABI.CurrentVersion` to 6.3 and ensure minor variances in JSON version numbers are permitted. (#1438)
This PR ensures that the JSON schema version 6.3 is considered
"supported" in Swift Testing 6.3. As well, it suppresses a thrown error
when the `version` field of a JSON record is slightly higher than the
expected schema. For example, if we ask Swift Testing the ABI version
corresponding to `VersionNumber(6, 3, 1)`, we'll get back `v6_3`, but
then if we try to decode a JSON record whose `version` field equals
`"6.3.1"`, we'll throw a version mismatch error. This error is now
suppressed unless the versions really do mismatch.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.

---------

Co-authored-by: Stuart Montgomery <smontgomery@apple.com>