Start adoption of `@section` and `@used`.
This PR begins our migration to `@section` and `@used`. Because of a bug in the Swift compiler (https://github.com/swiftlang/swift/issues/85411), we still need to rely on an unsupported/experimental language feature (CompileTimeValuesPreview). When that bug is resolved, we can remove checks for CompileTimeValuesPreview and fully "activate" this new code path.

Note that because of that bug, this PR does _not_ resolve #1371.