Include the API level in the Android OS version.
When a test run starts, we emit some metadata including the OS version for diagnostic purposes. This PR adds Android's API level (SDK version) to the OS version info we emit on that platform. This is an Android-specific concept and is used in Swift for `{#,@}available(Android)` expressions and attributes.